Added form validation and connected database
When employee data is submitted it will take to another route and add it to database and show a message "Employee Created".
Also, added validate.js to validate form data using regular expression.
If there are errors it will show in red in the page and will prevent the form from submitting.